backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pandas
numpy
matplotlib
lux-api
pandasai
ydata-profiling
openai
requests
python-multipart
tinydb
gunicorn
